// types/patient.ts
export type Patient = {
  id: string;
  first_name: string;
  last_name: string;
  address: string;
  phone: string;
  email: string;
  physician_id: string;
  physician_first_name: string;
  physician_last_name: string;
  physician_phone: string;
  physician_bill: string;
  prescription: string;
  dose: string;
  visit_date: string;
  next_visit: string;
};